refactor(tags): move TagSelectionDialog + CreateTagForm to components/tags
